requirements.txt
aiomysql==0.3.2
fastapi==0.135.1
httpx==0.28.1
pydantic==2.12.5
python_bcrypt==0.3.2
PyYAML==6.0.3
tldextract==5.3.1
uvicorn==0.42.0
